pyproject.toml
[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
python = "^3.13"
fastapi = {extras = ["all"], version = "^0.115.3"}
pydantic = {extras = ["email"], version = "^2.9.2"}
pydantic-settings = "^2.6.0"
icecream = "^2.1.3"
motor = "^3.6.0"
python-multipart = "^0.0.16"
pyjwt = "^2.9.0"
cryptography = "^43.0.3"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
gunicorn = "^23.0.0"
